DFT+U
A big modification of codes structure of DFT+U
